+++ src/main/kotlin/kondorcet/Ballot.kt
@@ -1,13 +1,15 @@
 package kondorcet
 
 /**
- * Represent a ballot
+ * Represent a vote ballot
+ *
+ * A ballot contains candidates ordered by preferences.
+ *
+ * A ballot cannot contains more that once each candidate
  */
 interface Ballot<out T : Any> {
 
-    /**
-     * Ordered list of candidates
-     */
+    /** Ordered list of candidates */
     val orderedCandidates: List<Set<T>>
 
     /** Candidates */
@@ -15,7 +17,7 @@ interface Ballot<out T : Any> {
         get() = orderedCandidates.flatten().toSet()
 
     /**
-     * List of winner (many candidates mean ex aequo)
+     * List of winner (many candidates means ex aequo)
      */
     val winners: Set<T>
         get() = orderedCandidates.firstOrNull() ?: emptySet()
@@ -25,9 +27,4 @@ interface Ballot<out T : Any> {
      */
     val winner: T?
         get() = winners.let { if (it.size == 1) it.first() else null }
-
-    /**
-     * Return true if, and only if, the ballot contains at least one candidate twice
-     */
-    fun hasDuplicates() = candidates.size != orderedCandidates.flatten().size
 }